feat: implement worker-side service handlers for node services and update related references
